refactor: refactor project

+++ src/main/java/com/mywallet/api/domain/entity/Wallet.java
@@ -36,14 +36,14 @@ public class Wallet implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "category", nullable = false)
     private WallateCategory category;
 
     @Column(name = "quantity")
-    private Integer quantity;
+    private Long quantity;
 
     @Column(name = "percent")
     private Integer percent;
