+++ src/main/java/com/mywallet/api/domain/model/AcessModel.java
@@ -3,5 +3,5 @@ package com.mywallet.api.domain.model;
 import lombok.Builder;
 
 @Builder
-public record AcessModel(Integer id, String name, String token) {
+public record AcessModel(Long id, String name, String token) {
 }
